[Fix] Fix wrong format string type specifier in EffectInfo::GetVolCmdParamInfo().
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@7103 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/EffectInfo.cpp
@@ -1015,7 +1015,7 @@ bool EffectInfo::GetVolCmdParamInfo(const ModCommand &m, LPSTR s) const
 				param = sndFile.GetSample(smp).cues[m.vol - 1];
 			else
 				param = m.vol << 11;
-			sprintf(s, "Cue %u: %u", m.vol, param);
+			sprintf(s, "Cue %u: %llu", m.vol, static_cast<unsigned long long>(param));
 		} else
 		{
 			strcpy(s, "continue");
